Change showMovie to showMovies: lists movie-type children of current content item. New query. A bit of error handling on server and in client. Constants set more DOM-economically in the part. Removed redundant OPTIONS_HEADERS.
