--- src/main/resources/site/parts/show-movie/show-movie.xml
@@ -1,5 +1,0 @@
-<part>
-    <display-name>Show movie</display-name>
-    <config>
-    </config>
-</part>
