+++ src/main/resources/site/parts/show-movies/show-movies.xml
@@ -1,0 +1,5 @@
+<part>
+    <display-name>Show movies</display-name>
+    <config>
+    </config>
+</part>
